Vocabulary

+++ src/main/java/com/example/HustLearning/repository/TopicRepository.java
@@ -3,5 +3,6 @@ package com.example.HustLearning.repository;
 import com.example.HustLearning.entity.Topic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopicRepository extends JpaRepository<Topic,Integer> {
+public interface TopicRepository extends JpaRepository<Topic, Long> {
+
 }
